added search input
