+++ src/index.css
@@ -2,4 +2,7 @@ body {
   margin: 0;
   max-width: 100%;
   overflow-x: hidden;
+  font-family: Helvetica Neue;
+  font-weight: 200;
+  letter-spacing: 1.92px;
 }
