+++ src/styles/header/header.scss
@@ -5,3 +5,48 @@
   width: 100%;
   flex-shrink: 0;
 }
+
+.navbar-header {
+  display: inline-flex;
+  align-items: flex-start;
+  gap: 20px;
+}
+
+img {
+  width: 20px;
+  height: 20px;
+}
+
+.headerLogo {
+  width: 32.934px;
+  height: 32px;
+  flex-shrink: 0;
+}
+
+.inputSearchFieldContainer {
+  display: flex;
+  justify-content: flex-start;
+  width: 334px;
+  height: 30px;
+  padding: 6px 12px;
+  align-items: center;
+  gap: 8px;
+  flex-shrink: 0;
+  border-radius: 4px;
+  border: 0.5px solid #000;
+  outline: none;
+
+  :focus {
+    outline: none;
+  }
+}
+
+.inputField {
+  border: none;
+  background: transparent;
+  margin-left: 30px;
+}
+
+.headerNavbarContainer {
+  border: 2px solid red;
+}
